requirements.txt
requests
beautifulsoup4
pyjwt
python-owasp-zap-v2.4